Refactor sending sync emails as a mixin
To ensure the synchronous send method is used when the mixin is used, it
has to inherit from EmailBase.
EmailBase had to be moved to a separate file to prevent circular
dependencies.
