+++ adhocracy4/emails/mixins.py
@@ -1,9 +1,10 @@
 from email.mime.image import MIMEImage
 
 from django.contrib.staticfiles import finders
+from .base import EmailBase
 
 
-class PlatformEmailMixin():
+class PlatformEmailMixin:
     """
     Attaches the static file images/logo.png so it can be used in an html
     email.
@@ -17,3 +18,12 @@ class PlatformEmailMixin():
             logo.add_header('Content-ID', '<{}>'.format('logo'))
             return attachments + [logo]
         return attachments
+
+
+class SyncEmailMixin(EmailBase):
+    """Send Emails synchronously."""
+
+    @classmethod
+    def send(cls, object, *args, **kwargs):
+        """Call dispatch immediately"""
+        return cls().dispatch(object, *args, **kwargs)
